chore: extract component
