+++ packages/workspace-fusion/src/lib/integrations/power-bi/utils/addPowerBi.tsx
@@ -1,17 +1,11 @@
 import { PowerBI, PowerBiController, IBasicFilter } from '@equinor/workspace-powerbi';
 import { Tab } from '@equinor/workspace-react';
-import { CircularProgress, Icon, Popover } from '@equinor/eds-core-react';
-import { info_circle } from '@equinor/eds-icons';
 import { PowerBiHeader } from '../components/workspaceHeader/PowerBiHeader';
 import { PowerBiIcon } from '../icons/PowerBiIcon';
 import { FilterConfig, PowerBiConfig } from '../';
-import { BaseEvent } from '@equinor/workspace-core';
 import { HeaderIcon, useWorkspaceHeaderComponents } from '../../../context';
-import { ReactNode, Suspense, useEffect, useRef, useState } from 'react';
-import { tokens } from '@equinor/eds-tokens';
-import { createPortal } from 'react-dom';
-import { ErrorBoundary } from 'react-error-boundary';
-Icon.add({ info_circle });
+import { useEffect } from 'react';
+import { MetaPopup } from '../components/PowerBiPopover';
 
 export function addPowerBi(powerBiConfig: PowerBiConfig | undefined): undefined | Tab {
 	if (!powerBiConfig) return;
@@ -73,40 +67,3 @@ const PowerBiWrapper = (powerBiConfig: PowerBiConfig & { controller: PowerBiCont
 		/>
 	);
 };
-
-type MetaPopupProps = {
-	children: ReactNode;
-	anchor: HTMLElement;
-};
-const MetaPopup = ({ children, anchor }: MetaPopupProps) => {
-	const [isOpen, setIsOpen] = useState(false);
-	const pRef = useRef<HTMLDivElement | null>(null);
-
-	return (
-		<>
-			<div ref={pRef}>
-				<Icon
-					color={tokens.colors.text.static_icons__tertiary.hex}
-					onClick={() => setIsOpen((s) => !s)}
-					name={info_circle.name}
-				/>
-			</div>
-			{isOpen &&
-				createPortal(
-					<Popover placement="bottom" anchorEl={pRef.current} open={isOpen}>
-						{/* TODO: Parse error */}
-						<ErrorBoundary FallbackComponent={() => <div>Failed to load report info</div>}>
-							<Suspense fallback={<Loading />}>{children}</Suspense>
-						</ErrorBoundary>
-					</Popover>,
-					anchor
-				)}
-		</>
-	);
-};
-
-const Loading = () => (
-	<div style={{ height: '400px', width: '300px', display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
-		<CircularProgress />
-	</div>
-);
